requirements.txt
# Turing Smart Screen Monitor - Dependencies
# Core communication
pyserial>=3.5

# Image generation and rendering
Pillow>=10.0.0

# System monitoring
psutil>=5.9.0

# GPU monitoring (optional - NVIDIA only)
GPUtil>=1.4.0

# System tray icon
pystray>=0.19.0
